+++ src/components/whenModal/WhenModal.module.css
@@ -8,52 +8,51 @@
   background: rgba(0, 0, 0, 0.6);
   display: none;
 }
-
 .modal {
   padding: 20px;
   background: white;
-  width: 1100px;
-  border-radius: 5px;
+  width: 1200px;
+  border-radius: 10px;
   box-shadow: 0 0 6px 0 rgb(0 0 0 / 50%);
   height: 100%;
   overflow: auto;
   margin: auto;
 }
+
+/* title section */
 .titleWrapper {
   display: flex;
   justify-content: center;
   align-items: center;
 }
-
 .modalTitle {
-  width: 900px;
+  width: 1000px;
   display: flex;
   justify-content: space-between;
-  padding: 70px 0 20px 0;
+  padding: 90px 0 20px 0;
   border-bottom: 1px solid gray;
 }
-.inputLocation {
-  display: block;
-  margin: 20px auto;
+.whereText {
+  font-weight: bold;
+  font-size: 35px;
 }
-.location {
+.closeModal {
+  cursor: pointer;
+}
+.calendarWrapper {
   display: flex;
   justify-content: center;
+  margin: 0 100px;
+  position: relative;
 }
-.locationWrapper {
-  width: 400px;
+.prevBtn {
+  position: absolute;
+  left: 0%;
+  cursor: pointer;
 }
-.domestic {
-  padding-left: 20px;
-}
-.cities {
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-}
-.city {
-  padding: 20px;
-}
-.space {
-  height: 1000px;
+
+.nextBtn {
+  position: absolute;
+  right: 0%;
+  cursor: pointer;
 }
